Previous results:

```
BM_MatchedSingleQuery              99 ns         99 ns    6865572
BM_MatchedFullQuery               321 ns        320 ns    2206163
BM_UnmatchedSingleQuery           308 ns        308 ns    2270457
BM_UnmatchedFullQuery             313 ns        312 ns    2218307
BM_OneVariableFullMatch           449 ns        448 ns    1552719
BM_OneVariableMatchUnmatch        769 ns        768 ns     889476
```

New results:

```
BM_MatchedSingleQuery              60 ns         60 ns   11614595
BM_MatchedFullQuery               314 ns        314 ns    2242016
BM_UnmatchedSingleQuery            56 ns         56 ns   12320256
BM_UnmatchedFullQuery             272 ns        272 ns    2609243
BM_OneVariableFullMatch           392 ns        392 ns    1783676
BM_OneVariableMatchUnmatch        476 ns        476 ns    1467029
```